+++ addons/binding/org.openhab.binding.networkcameramotiondetection/ESH-INF/binding/binding.xml
@@ -11,7 +11,7 @@
 		<parameter name="port" type="integer" required="true" min="1" max="65335">
 			<label>TCP Port</label>
 			<description>TCP port of the FTP server</description>
-			<default>2000</default>
+			<default>2121</default>
 		</parameter>
 		<parameter name="idleTimeout" type="integer" required="true" min="0" max="65335">
             <label>Idle timeout</label>
