+++ addons/binding/org.openhab.binding.networkcameramotiondetection/META-INF/MANIFEST.MF
@@ -13,6 +13,7 @@ Import-Package:
  com.google.common.io,
  org.apache.commons.io,
  org.apache.commons.lang,
+ org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.core.library.types,
  org.eclipse.smarthome.core.thing,
