+++ addons/binding/org.openhab.binding.networkcameramotiondetection/OSGI-INF/NetworkCameraMotionDetectionHandlerFactory.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2014-2015 openHAB UG (haftungsbeschraenkt) and others.
+    Copyright (c) 2014-2017 openHAB UG (haftungsbeschraenkt) and others.
     All rights reserved. This program and the accompanying materials
     are made available under the terms of the Eclipse Public License v1.0
     which accompanies this distribution, and is available at
@@ -16,6 +16,6 @@
       <provide interface="org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory"/>
    </service>
    
-   <property name="service.pid" type="String" value="org.openhab.networkcameramotiondetection2" />
+   <property name="service.pid" type="String" value="org.openhab.networkcameramotiondetection" />
 
 </scr:component>
